+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 CHANGELOG inspiration from http://keepachangelog.com/.
 
 ## Unreleased
+* Update influxdb_queries.yml and LoadQueries for InfluxDB 0.10
 
 ## [0.4.1] - March 4, 2016
 * Skip loading continuous queries if influxdb_queries.yml file doesn't exist.